cloudfunctions/hero-data/requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
requests==2.22.0

